+++ cms/api/base.py
@@ -1,5 +1,0 @@
-from cms.views.base import BaseCmsView
-
-
-class ApiBase(BaseCmsView):
-    pass